refactor: sort-highlight-column -> sortColumn, highlightColumn modules
