+++ public/js/table/build-dom.js
@@ -3,7 +3,7 @@ import { collectRowsData, collectTableDataAndSave, confirmDeletingTable } from '
 import buildTables from './build-tables.js';
 import { addColumn, deleteColumn } from './add-delete-column.js';
 import { addRow, deleteRow} from './add-delete-row.js';
-import { sortColumn, highlightColumn } from './sort-highlight-column.js';
+import { sortColumn, highlightColumn } from './table-utils/modifiers/index.js';
 import watch from '../utils/watch.js';
 import { editTableTitle, changeColumnsWidth, resetStyles } from './table-utils/modifiers/index.js';
 import closeTable from './table-utils/close-table.js';
