+++ public/js/table/table-utils/modifiers/index.js
@@ -5,6 +5,8 @@ import enactShowHideResizer from './enact-show-hide-resizer.js';
 import removeEmptyColumns from './remove-empty-columns.js';
 import removeLastEmptyRows from './remove-last-empty-rows.js';
 import resetStyles from './reset-styles.js';
+import sortColumn from './sort-column.js';
+import highlightColumn from './highlight-column.js';
 
 export {
   alignTextAreasHeight,
@@ -14,4 +16,6 @@ export {
   removeEmptyColumns,
   removeLastEmptyRows,
   resetStyles,
+  sortColumn,
+  highlightColumn,
 }
